backend/requirements.txt
# ─── Web Framework ───────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9       # required for file uploads

# ─── AI / LLM Layer ──────────────────────────────────────────────────────────
langchain==0.2.1
langchain-openai==0.1.8
langchain-community==0.2.1
openai==1.30.1

# ─── Document Parsers ────────────────────────────────────────────────────────
pypdf==4.2.0                  # PDF parsing
python-docx==1.1.2            # DOCX parsing

# ─── Database ────────────────────────────────────────────────────────────────
asyncpg==0.29.0               # async PostgreSQL driver
sqlalchemy[asyncio]==2.0.30   # ORM with async support
pgvector==0.3.2               # pgvector Python client
alembic==1.13.1               # database migrations

# ─── Config & Utilities ──────────────────────────────────────────────────────
pydantic==2.7.1               # data validation (FastAPI uses this internally)
pydantic-settings==2.3.0      # .env file loading
python-dotenv==1.0.1
tiktoken==0.7.0               # token counting for OpenAI models
